+++ src/components/store/sub-component/StoreTimeSet.tsx
@@ -1,9 +1,9 @@
 import { useGetQuery } from '@/hooks/store/useGetQuery';
 import { Fragment, useEffect, useState } from 'react';
-import { convertTimeFormat, timeSet } from '../utils/time-format';
+import { convertTimeFormat, timeSet } from '../../utils/time-format';
+import styles from '../styles/StroeContents.module.css';
 import StoreSelectBox from './StoreSelectBox';
 import StoreSetButton from './StoreSetButton';
-import styles from './styles/StroeContents.module.css';
 
 const StoreTimeSet = ({ userId }: { userId: string }) => {
   const { data } = useGetQuery({ userId });
